Create Contract2.sol
In the Movie struct, it might be better to store the productionCo address as the first field to allow for easier access and usage.
In the addMovie function, the msg.sender address should be stored as the first field when creating a new Movie struct.
In the getMovie function, the productionCo address field is missing from the return statement.
In the getMovies function, it is not clear what should be returned. Perhaps this function should return an array of all the movie IDs, or the total number of movies in the store.
Here is the updated code